mei: add read callback on demand for fixed_address clients
Upstream-Status: Submitted[https://patchwork.kernel.org/patch/9181209/]

The Fixed address clients do not work with the flow control, and the
packet RX callback was allocated upon TX with anticipation of a
following RX. This won't work if the clients with unsolicited Rx. Rather
than preparing read callback upon a write we allocate one directly on
the reciev path if one doesn't exists.

Change-Id: I872f261cfed92832e9930fe32581c893e2eadab0
Signed-off-by: Alexander Usyskin <alexander.usyskin@intel.com>
Signed-off-by: Tomas Winkler <tomas.winkler@intel.com>