lx-dialig: fix merge issues
The patch [menuconfig,check-lxdiaglog.sh: Allow specification of ncurses
location] has been carried since the 3.14 kernel. As it has been updated
multiple times, the if block checking for CROSS_CURSES_LIB has migrated
to an invalid location in the checks.

This change moves it back on top where it belongs.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>