drm/i915: Fix v4.4.20 -> v4.4.22 i915 merge.
Some patches did not merge cleanly in previous merge, causing build
failures. The i915 changes between 94e5bb30ea and 1fd9c1008d were
reverted and the patches were reapplied as follows, then the results
were squashed.

Commit a554bd7ccf "drm/i915: Fix hpd live status bits for g4x" was
already present before the merge and did not need reapplied or fixing.

Commit af115b1a3a "drm/i915: Only ignore eDP ports that are connected"
was resolved by applying the upstream version of the patch instead of
the version that came from stable. This applied cleanly.

Commit e153f52df9 "drm/i915: Check VBT for port presence in
addition to the strap on VLV/CHV" did not apply cleanly with either the
upstream or stable versions, and had to be fixed manually.

Commit 0da2f8261e "drm/i915: Avoid pointer arithmetic in calculating
plane surface offset" was already present before the merge and did not
need reapplied or fixing.

Signed-off-by: California Sullivan <california.l.sullivan@intel.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>